+++ vimiv/api/settings.py
@@ -470,6 +470,11 @@ class thumbnail:  # pylint: disable=invalid-name
     """Namespace for thumbnail related settings."""
 
     size = ThumbnailSizeSetting("thumbnail.size", 128, desc="Size of thumbnails")
+    save = BoolSetting(
+        "thumbnail.save",
+        True,
+        desc="Save new thumbnails to the disk in the shared icon cache for later use"
+    )
 
 
 class slideshow:  # pylint: disable=invalid-name