+++ docs/changelog.rst
@@ -3,6 +3,10 @@ Changelog
 
 All notable changes to vimiv are documented in this file.
 
+v0.10.0 (unreleased)
+--------------------
+
+
 v0.9.0 (2023-07-15)
 -------------------